requirements.txt
# --- Web API ---
fastapi>=0.110
uvicorn[standard]>=0.23

# --- HTTP / Validation ---
requests>=2.31
pydantic>=2,<3

# --- LangChain & Embeddings ---
nltk==3.9.1
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
transformers
huggingface-hub

# --- ML Backend ---
torch
numpy==1.26.4

# --- Utilities ---
attrs
pybind11>=2.10

# --- Database (PostgreSQL via SQLAlchemy) ---
SQLAlchemy>=2,<3
psycopg2-binary==2.9.9
# psycopg2==2.9.9

# --- NLP ---
fasttext-wheel==0.9.2
# fasttext==0.9.2